requirements.txt
python-telegram-bot[webhooks]==20.3
openai>=1.0.0
gspread
oauth2client
